Updated project to Xcode 7, j2objc 0.9.8.1.

+++ Classes/J2RBoard.h
@@ -32,6 +32,6 @@
 
 - (void)update;
 
-- (void)updateSquare:(NSInteger)column withRow:(NSInteger)row;
+- (void)updateSquare:(jint)column withRow:(jint)row;
 
 @end
